+++ src/main/java/tornadofx/Binding.kt
@@ -25,34 +25,34 @@ import java.time.LocalTime
 import java.util.*
 import java.util.concurrent.Callable
 
-inline fun <reified T> ComboBoxBase<T>.bind(property: Property<T>, readonly: Boolean = false) =
-        if (readonly) valueProperty().bind(property) else valueProperty().bindBidirectional(property)
+inline fun <reified T> ComboBoxBase<T>.bind(property: ObservableValue<T>, readonly: Boolean = false) =
+        if (readonly || (property !is Property<*>)) valueProperty().bind(property) else valueProperty().bindBidirectional(property as Property<T>)
 
-fun DatePicker.bind(property: Property<LocalDate>, readonly: Boolean = false) =
-        if (readonly) valueProperty().bind(property) else valueProperty().bindBidirectional(property)
+fun DatePicker.bind(property: ObservableValue<LocalDate>, readonly: Boolean = false) =
+        if (readonly || (property !is Property<*>)) valueProperty().bind(property) else valueProperty().bindBidirectional(property as Property<LocalDate>)
 
-fun ProgressIndicator.bind(property: Property<Number>, readonly: Boolean = false) =
-        if (readonly) progressProperty().bind(property) else progressProperty().bindBidirectional(property)
+fun ProgressIndicator.bind(property: ObservableValue<Number>, readonly: Boolean = false) =
+        if (readonly || (property !is Property<*>)) progressProperty().bind(property) else progressProperty().bindBidirectional(property as Property<Number>)
 
-inline fun <reified T> ChoiceBox<T>.bind(property: Property<T>, readonly: Boolean = false) =
-        if (readonly) valueProperty().bind(property) else valueProperty().bindBidirectional(property)
+inline fun <reified T> ChoiceBox<T>.bind(property: ObservableValue<T>, readonly: Boolean = false) =
+        if (readonly || (property !is Property<*>)) valueProperty().bind(property) else valueProperty().bindBidirectional(property as Property<T>)
 
-fun CheckBox.bind(property: Property<Boolean>, readonly: Boolean = false) =
-        if (readonly) selectedProperty().bind(property) else selectedProperty().bindBidirectional(property)
+fun CheckBox.bind(property: ObservableValue<Boolean>, readonly: Boolean = false) =
+        if (readonly || (property !is Property<*>)) selectedProperty().bind(property) else selectedProperty().bindBidirectional(property as Property<Boolean>)
 
-fun Slider.bind(property: Property<Number>, readonly: Boolean = false) =
-        if (readonly) valueProperty().bind(property) else valueProperty().bindBidirectional(property)
+fun Slider.bind(property: ObservableValue<Number>, readonly: Boolean = false) =
+        if (readonly || (property !is Property<*>)) valueProperty().bind(property) else valueProperty().bindBidirectional(property as Property<Number>)
 
 inline fun <reified T : Any> Labeled.bind(property: ObservableValue<T>, readonly: Boolean = false, converter: StringConverter<T>? = null, format: Format? = null) =
         bindStringProperty(textProperty(), converter, format, property, readonly)
 
-inline fun <reified T : Any> TitledPane.bind(property: Property<T>, readonly: Boolean = false, converter: StringConverter<T>? = null, format: Format? = null) =
+inline fun <reified T : Any> TitledPane.bind(property: ObservableValue<T>, readonly: Boolean = false, converter: StringConverter<T>? = null, format: Format? = null) =
         bindStringProperty(textProperty(), converter, format, property, readonly)
 
-inline fun <reified T : Any> Text.bind(property: Property<T>, readonly: Boolean = false, converter: StringConverter<T>? = null, format: Format? = null) =
+inline fun <reified T : Any> Text.bind(property: ObservableValue<T>, readonly: Boolean = false, converter: StringConverter<T>? = null, format: Format? = null) =
         bindStringProperty(textProperty(), converter, format, property, readonly)
 
-inline fun <reified T : Any> TextInputControl.bind(property: Property<T>, readonly: Boolean = false, converter: StringConverter<T>? = null, format: Format? = null) =
+inline fun <reified T : Any> TextInputControl.bind(property: ObservableValue<T>, readonly: Boolean = false, converter: StringConverter<T>? = null, format: Format? = null) =
         bindStringProperty(textProperty(), converter, format, property, readonly)
 
 inline fun <reified T : Any> bindStringProperty(stringProperty: StringProperty, converter: StringConverter<T>?, format: Format?, property: ObservableValue<T>, readonly: Boolean) {